--- a/include/GafferSceneUI/StandardLightVisualiser.h
+++ b/include/GafferSceneUI/StandardLightVisualiser.h
@@ -106,7 +106,7 @@ class GAFFERSCENEUI_API StandardLightVisualiser : public IECoreGLPreview::LightV
 		static IECoreGL::ConstRenderablePtr environmentSphereSurface( IECore::ConstDataPtr textureData, int textureMaxResolution, const Imath::Color3f &fallbackColor  );
 
 		// Spread is generally rendered as an angle, rather than in light space,
-		// as such, this should generally only be used as an ornament type visualisation.
+		// as such, this should generally only be used as a uniformly scaled visualisation.
 		static IECoreGL::ConstRenderablePtr areaSpread( float spread );
 
 		static IECoreGL::ConstRenderablePtr quadWireframe( const Imath::V2f &size );