--- a/python/GafferSceneTest/IECoreGLPreviewTest/__init__.py
+++ b/python/GafferSceneTest/IECoreGLPreviewTest/__init__.py
@@ -35,6 +35,7 @@
 ##########################################################################
 
 from RendererTest import RendererTest
+from VisualiserTest import VisualiserTest
 
 if __name__ == "__main__":
 	import unittest