--- a/python/GafferSceneUI/CameraUI.py
+++ b/python/GafferSceneUI/CameraUI.py
@@ -350,7 +350,7 @@ plugsMetadata = {
 
 	],
 
-	"visualiserAttributes.ornamentScale" : [
+	"visualiserAttributes.scale" : [
 
 			"description",
 			"""