… Changes.md
… python/GafferSceneUITest/VisualiserTest.py
… src/GafferArnoldUI/LightBlockerVisualiser.cpp
… src/GafferScene/Camera.cpp
… src/GafferScene/IECoreGLPreview/AttributeVisualiser.cpp
… src/GafferScene/IECoreGLPreview/LightFilterVisualiser.cpp
… src/GafferScene/IECoreGLPreview/LightVisualiser.cpp
… src/GafferScene/IECoreGLPreview/Renderer.cpp
… src/GafferScene/IECoreGLPreview/Visualiser.cpp
… src/GafferScene/Light.cpp
… src/GafferScene/OpenGLAttributes.cpp
… src/GafferSceneModule/IECoreGLPreviewBinding.cpp
… src/GafferSceneUI/CameraVisualiser.cpp
… src/GafferSceneUI/ClippingPlaneVisualiser.cpp
… src/GafferSceneUI/CoordinateSystemVisualiser.cpp
… src/GafferSceneUI/ProceduralVisualiser.cpp
… src/GafferSceneUI/SceneView.cpp
… src/GafferSceneUI/StandardLightVisualiser.cpp
… src/GafferVDBUI/VDBVisualiser.cpp
… startup/GafferScene/lightCompatibility.py