Supersedes #3551 and #3056. Sorry there is a bunch in here now, but it was hard to do this refactor and still maintain two independent PRs for the spot/camera frustum work.

In practice, as we developed more visualisers and attempted to match behaviour across renderers, it became clear that the concrete-type approach we went for in #3514 wouldn't scale to all the permutations required. Particularly as specific rendering behaviours don't match to specific user-land categories such as 'frustums'. The requirement to expose drawing toggles in the UI made the typed approach untenable.

This PR switches to an alternate approach, refactoring Visualisers such that they return a vector of `Visualisation` structs instead of a type-keyed array.

We experimented with various combinations flags, and/or enums, and this ended up with (hopefully) the most logical naming conventions, visualiser and renderer logic.

Improvements
--------------

- Viewer : 
  - Added menu items to control frustum visualisations. These can also be controlled via the `OpenGLAttributes` node (#3569).
  - Added spotlight light cone visualisation (#3569).
  - Added visualisation of camera frustums (#3569).
  - Improved <kbd>f</kbd> framing behaviour for light visualisations (#3569).
- Camera : Added Visualisation controls to the Camera node to allow frustum visualisation to be easily overridden per-camera (#3569).

API
---

- IECoreGLPreview : Added support for more flexible visualisations via the `Visualisation` struct (#3569).

Breaking Changes
------------------

 - IECoreGLPreview : Changed the return type of `(Light|LightFilter|Attribute|Object)Visualiser` classes. Visualisations are now a vector of `Visualisation` structs (#3569).


## Cameras

![image](https://user-images.githubusercontent.com/896779/72168089-5e7bce00-33c4-11ea-87b4-e377bf863411.png)

## Spot Lights

![image](https://user-images.githubusercontent.com/896779/72148326-6887d780-3398-11ea-8261-42cbc3077ae8.png)