@themissingcow 2020-01-17
Need to add some tests yet.