@themissingcow 2020-01-20
Tests added.