@themissingcow 2020-01-30
> I haven't done any performance testing on this yet - I don't suppose you have? From what I can see, StandardLightVisualiser always returns 4 visualisations now, even though many will be empty. 

Good spot, thanks. Fixed this in https://github.com/GafferHQ/gaffer/pull/3569/commits/866562dc4abeb37de1546e005668fd90214389b4, and adjusted subsequent additions.

> I'm not sure how that is going to play with the "loop over N visualisers M times" approach the Renderer is taking, and if perhaps we're making the wrong trade-off there in an attempt to minimise pushes. I'll have a play with it tomorrow.

Will do performance/regression testing today and post some comparisons.