@johnhaddon 2020-02-03
Thanks Tom, LGTM. Let's get some performance numbers and then merge if all is well.