@themissingcow 2020-02-03
Quick manual performance tests, decorating `ViewportGadget.draw`. 

Set up is 625 objects.

![image](https://user-images.githubusercontent.com/896779/73676997-48a0b680-46ad-11ea-8948-748c707cc314.png)

## Spotlight (with gobo, decay and barn door visualisations)
### 55
102 draw calls 81.7ms average
98 draw calls 85.9ms average
98 draw calls 86.4ms average
### master
107 draw calls 78.0ms average
103 draw calls 80.6ms average
103 draw calls 81.5ms average
### PR
85 draw calls 98.6ms average
88 draw calls 95.9ms average
87 draw calls 97.9ms average
### PR (no frustum)
97 draw calls 87.1ms average
96 draw calls 87.9ms average
95 draw calls 89.5ms average

## Quads
### 55
173 draw calls 39.8ms average
185 draw calls 38.0ms average
175 draw calls 39.1ms average
### master
169 draw calls 42.5ms average
172 draw calls 42.5ms average
174 draw calls 42.2ms average
### PR
163 draw calls 45.1ms average
164 draw calls 45.0ms average
164 draw calls 45.6ms average

## Cubes
### 55
327 draw calls 12.1ms average
283 draw calls 12.9ms average
321 draw calls 12.5ms average
### master
307 draw calls 13.2ms average
313 draw calls 13.4ms average
326 draw calls 13.3ms average
### PR
317 draw calls 13.1ms average
324 draw calls 13.2ms average
325 draw calls 13.4ms average