@johnhaddon 2020-02-04
Thanks Tom, let's deal with performance tweaks in a followup, so we can get a preview release out to folks. Could you rebase to fix the conflict and squash in the remaining fixup commits?